% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{clip_footnotes}
\alias{clip_footnotes}
\title{Remove the footnotes from the bottom of the page}
\usage{
clip_footnotes(page)
}
\arguments{
\item{page}{a page of text generated by \code{pdf_text()}}
}
\value{
a page of text with no footnote text
}
\description{
Remove the footnotes from the bottom of the page
}
