% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{extract_footnotes}
\alias{extract_footnotes}
\title{Extract footnote text}
\usage{
extract_footnotes(page)
}
\arguments{
\item{page}{a page of text generated by \code{pdf_text()}}
}
\value{
a list of footnotes in the page and their indices
}
\description{
Extract footnote text
}
