% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-text.R
\name{extract_sidebar_text}
\alias{extract_sidebar_text}
\title{extract sidebar text from page}
\usage{
extract_sidebar_text(page)
}
\arguments{
\item{page}{}
}
\description{
extract sidebar text from page
}
