% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{markdownify_footnotes}
\alias{markdownify_footnotes}
\title{replace footnotes inline with markdown style footnote indicators.
Exclude the words fig and any month abbreviations}
\usage{
markdownify_footnotes(page)
}
\arguments{
\item{page}{a page of text generated by \code{pdf_text()}}
}
\value{
a page with all inline footnotes replaced with markdown style footnotes
}
\description{
replace footnotes inline with markdown style footnote indicators.
Exclude the words fig and any month abbreviations
}
