% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplates.R
\name{remove_contact_page}
\alias{remove_contact_page}
\title{a page with instructions for contacting gao appears at the end of
most reports. This function detects and removes that page by searching
for the GAO phone number, which should not appear in the body of
any report}
\usage{
remove_contact_page(text)
}
\arguments{
\item{text}{the text of a report as extracted by \code{pdf_text()}}
}
\description{
a page with instructions for contacting gao appears at the end of
most reports. This function detects and removes that page by searching
for the GAO phone number, which should not appear in the body of
any report
}
